+++ libft/ft_strlen.c
@@ -6,16 +6,16 @@
 /*   By: jde-mour <jerome@demourgues.com>           +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/05 11:41:11 by jde-mour     #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/17 17:49:28 by jde-mour    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/11/13 17:04:51 by jde-mour    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
